rename bloodhound_pki_template to bloodhound_certtemplate
