+++ bofhound/ad/models/__init__.py
@@ -12,6 +12,6 @@ from .bloodhound_rootca import BloodHoundRootCA
 from .bloodhound_aiaca import BloodHoundAIACA
 from .bloodhound_ntauthstore import BloodHoundNTAuthStore
 from .bloodhound_issuancepolicy import BloodHoundIssuancePolicy
-from .bloodhound_pki_template import BloodHoundCertTemplate
+from .bloodhound_certtemplate import BloodHoundCertTemplate
 from .bloodhound_domaintrust import BloodHoundDomainTrust
 from .bloodhound_crossref import BloodHoundCrossRef
