+++ bofhound/ad/models/bloodhound_certtemplate.py
